+++ include/sdf/JointAxis.hh
@@ -18,6 +18,7 @@
 #define SDF_JOINTAXIS_HH_
 
 #include <string>
+#include <ignition/math/Vector3.hh>
 #include "sdf/Element.hh"
 #include "sdf/Types.hh"
 #include "sdf/sdf_config.h"
@@ -98,14 +99,16 @@ namespace sdf
     /// \return True to interpret the axis xyz value in the parent model
     /// frame, false to use the joint frame.
     /// \sa void SetUseParentModelFrame(const bool _parentModelFrame)
-    public: bool UseParentModelFrame() const;
+    public: bool UseParentModelFrame() const
+        SDF_DEPRECATED(9.0);
 
     /// \brief Set whether to interpret the axis xyz value in the parent model
     /// instead of the joint frame.
     /// \param[in] _parentModelFrame True to interpret the axis xyz value in
     /// the parent model frame, false to use the joint frame.
     /// \sa bool UseParentModelFrame() const
-    public: void SetUseParentModelFrame(const bool _parentModelFrame);
+    public: void SetUseParentModelFrame(const bool _parentModelFrame)
+        SDF_DEPRECATED(9.0);
 
     /// \brief Get the physical velocity dependent viscous damping coefficient
     /// of the joint axis. The default value is zero (0.0).
@@ -225,6 +228,18 @@ namespace sdf
     /// \param[in] _dissipation The joint stop dissipation.
     /// \sa double Dissipation() const
     public: void SetDissipation(const double _dissipation) const;
+
+    /// Get the name of the coordinate frame in which this joint axis's
+    /// unit vector is expressed. An empty value implies the parent (joint)
+    /// frame.
+    /// \return The name of the xyz expressed-in frame.
+    public: const std::string& XyzExpressedIn() const;
+
+    /// Set the name of the coordinate frame in which this joint axis's
+    /// unit vector is expressed. An empty value implies the parent (joint)
+    /// frame.
+    /// \param[in] The name of the xyz expressed-in frame.
+    public: void SetXyzExpressedIn(const std::string &_frame);
 
     /// \brief Get a pointer to the SDF element that was used during
     /// load.